Add certs.py
Now cert generation is handled in its own utility module
