+++ models/proxy.py
@@ -5,11 +5,10 @@ import sys
 import itertools
 import threading
 
-from util import parser
+from util import parser, certs
 from util.logging_conf import logger
 from controllers import queue_manager
 from threading import Thread
-from OpenSSL import crypto
 
 
 class Server(Thread):
@@ -131,110 +130,32 @@ class Server(Thread):
     def join_with_script_dir(path):
         return os.path.join(os.path.dirname(os.path.abspath(__name__)), path)
 
-    @staticmethod
-    def generate_keypair(path=None):
-        key = crypto.PKey()
-        key.generate_key(crypto.TYPE_RSA, 2048)
-        if path:
-            with open(path, 'w+') as key_file:
-                key_file.write(crypto.dump_privatekey(crypto.FILETYPE_PEM, key).decode("utf-8"))
-        return key
-
-    @staticmethod
-    def generate_csr(hostname, key, path=None):
-        """
-        :param hostname: Subject root hostname to use when adding SANs
-        :param key: Subject's private key
-        :param path: Optional path for csr request output
-        :return:
-        """
-
-        san_list = [f"DNS.1:*.{hostname}",
-                    f"DNS.2:{hostname}"]
-
-        csr = crypto.X509Req()
-        csr.get_subject().CN = hostname
-        # SANs are required by modern browsers, so we add them
-        csr.add_extensions([
-            crypto.X509Extension(b"subjectAltName", False, ', '.join(san_list).encode())
-        ])
-        csr.set_pubkey(key)
-        csr.sign(key, "sha256")
-
-        if path:
-            with open(path, 'w+') as csr_file:
-                csr_file.write(crypto.dump_certificate_request(crypto.FILETYPE_PEM, csr).decode("utf-8"))
-        return csr
-
-    def generate_certificate(self, hostname: str):
-        # ref: https://stackoverflow.com/questions/10175812/how-to-generate-a-self-signed-ssl-certificate-using-openssl
-
-        host_cert_path = f"{self.certs_path}generated\\{hostname}"
-        key_file_path = f"{host_cert_path}\\{hostname}.key"
-        csr_file_path = f"{host_cert_path}\\{hostname}.csr"
-        cert_file_path = f"{host_cert_path}\\{hostname}.pem"
-
-        if not os.path.isdir(host_cert_path):
-            os.mkdir(host_cert_path)
-
-        root_ca_cert = crypto.load_certificate(crypto.FILETYPE_PEM, open(self.cacert, 'rb').read())
-        root_ca_key = crypto.load_privatekey(crypto.FILETYPE_PEM, open(self.cakey, 'rb').read())
-
-        print(f"{root_ca_cert} {root_ca_key}")
-
-        key = self.generate_keypair(key_file_path)
-        csr = self.generate_csr(hostname, key, csr_file_path)
-
-        # Generate cert
-
-        cert = crypto.X509()
-        cert.get_subject().CN = hostname
-        cert.set_serial_number(int.from_bytes(os.urandom(16), "big") >> 1)
-        cert.gmtime_adj_notBefore(0)
-        cert.gmtime_adj_notAfter(31536000)  # 1 year
-
-        # Yes we must add the SANs to the cert as well
-        san_list = [f"DNS.1:*.{hostname}",
-                    f"DNS.2:{hostname}"]
-
-        cert.add_extensions([
-            crypto.X509Extension(b"subjectAltName", False, ', '.join(san_list).encode())
-        ])
-
-        # Sign it
-        cert.set_issuer(root_ca_cert.get_subject())
-        cert.set_pubkey(csr.get_pubkey())
-
-        cert.sign(root_ca_key, 'sha256')
-
-        with open(cert_file_path, 'w+') as cert_file:
-            cert_file.write(crypto.dump_certificate(crypto.FILETYPE_PEM, cert).decode("utf-8"))
-
-        return cert_file_path, key_file_path
-
-    def relay_data(self, remote_socket, client_socket, client_data, chunk, port):
+    def relay_data(self, remote_socket, client_socket, client_data, port):
         # TODO: port only for debug rn, delete later
 
         _data = ''
         _chunk = ''
         while True:
-            _data = _data + client_data.decode('utf-8', errors='ignore')
+            try:
+                _data = _data + client_data.decode('utf-8', errors='ignore')
 
-            print(f"Client:\n{'=' * 200}\n{_data}\n{'=' * 200}")
+                print(f"Client:\n{'=' * 200}\n{_data}\n{'=' * 200}")
 
-            remote_socket.sendall(client_data)
+                remote_socket.sendall(client_data)
 
-            chunk = remote_socket.recv(self.buffer_size)
-            if not chunk:
-                break
+                chunk = remote_socket.recv(self.buffer_size)
+                if not chunk:
+                    break
 
-            # For debug
-            # TODO: calc buff len at beginning of handshake
-            _chunk = _chunk + chunk.decode('utf-8', errors='ignore')
+                # For debug
+                # TODO: calc buff len at beginning of handshake
+                _chunk = _chunk + chunk.decode('utf-8', errors='ignore')
 
-            print(f"Remote:\n{'=' * 200}\n{_chunk}\n{'=' * 200}")
+                print(f"Remote:\n{'=' * 200}\n{_chunk}\n{'=' * 200}")
 
-            client_socket.send(chunk)  # send back to browser
+                client_socket.send(chunk)  # send back to browser
+            except socket.error as error:
+                logger.error(f"ERROR: Unable to relay data {error}")
 
     def intercept(self, method, hostname, port):
         if self.intercepting:
@@ -247,7 +168,10 @@ class Server(Thread):
                     return
                 else:
                     logger.debug("genning cert")
-                    cert_path, key_path = self.generate_certificate(hostname)
+                    cert_path, key_path = certs.generate_certificate(self.certs_path,
+                                                                     hostname,
+                                                                     self.cacert,
+                                                                     self.cakey)
                     logger.debug(f"GOT certs: {cert_path, key_path}")
                     client_ssl_socket = self.wrap_client_socket(self.client_socket, cert_path, key_path)
                     logger.debug(f"GOT socket ssl client: {client_ssl_socket}")
@@ -281,15 +205,17 @@ class Server(Thread):
             remote_socket = socket.create_connection((hostname, port))
 
             if port == 80:
-                chunk = None
+
                 threading.Thread(target=self.relay_data, args=(remote_socket,
                                                                self.client_socket,
                                                                data,
-                                                               chunk,
                                                                port)).start()
             else:
 
-                cert_path, key_path = self.generate_certificate(hostname)
+                cert_path, key_path = certs.generate_certificate(self.certs_path,
+                                                                 hostname,
+                                                                 self.cacert,
+                                                                 self.cakey)
 
                 print(f"cert:{cert_path}\nkey{key_path}")
 
@@ -298,14 +224,11 @@ class Server(Thread):
                 client_ssl_socket = self.wrap_client_socket(self.client_socket, cert_path, key_path)
                 remote_ssl_socket = self.wrap_remote_socket(remote_socket, hostname)
 
-                chunk = None
-
                 ssl_client_data = client_ssl_socket.recv(4096)
 
                 threading.Thread(target=self.relay_data, args=(remote_ssl_socket,
                                                                client_ssl_socket,
                                                                ssl_client_data,
-                                                               chunk,
                                                                port)).start()
 
                 # remote_socket.close()
